Resolve issue #10 - [x] Change package name so it will include project name - [x] For NBSParser add additional fields: original author, description etc - [x] Make NBSParser a static util class - [x] Rename NBSParser to NBSFile - [x] Test serialization/deserializion of NBSSong - [x] In Spigot, package-info is outside the actual package - [x] Move All Song impls to new ``adventure`` module - [x] Rename standalone module to ``shared`` or ``common`` - [x] Move SongQueue to ``adventure`` - [x] Move InstrumentManager to Adventure
