+++ adventure/src/main/java/com/github/idankoblik/NoteInstrument.java
