+++ adventure/src/test/java/com/github/idankoblik/AbstractSpigotTest.java
