+++ adventure/src/test/java/com/github/idankoblik/manager/InstrumentManagerTest.java
