+++ build.gradle.kts
@@ -59,10 +59,17 @@ allprojects {
             name = "ApartiumNexus"
             url = uri("https://nexus.voigon.dev/repository/apartium/")
         }
+
+        maven {
+            name = "apartium-releases"
+            url = uri("https://nexus.voigon.dev/repository/apartium-releases")
+        }
     }
 
     dependencies {
+        implementation("net.apartium.cocoa-beans:common:${findProperty("cocoabeans.version")}")
         implementation("org.jetbrains:annotations:${findProperty("jetbrains.version")}")
+        compileOnlyApi("com.fasterxml.jackson.core:jackson-annotations:${findProperty("jackson.annotations.version")}")
 
         testImplementation(platform("org.junit:junit-bom:${findProperty("junit.version")}"))
         testImplementation("org.junit.jupiter:junit-jupiter")
