+++ gradle.properties
@@ -4,3 +4,5 @@ spigot.version=1.8.8-R0.1-SNAPSHOT
 kyori.bukkit.version=4.3.1
 kyori.version=4.17.0
 mock.bukkit.version=3.9.0
+jackson.annotations.version = 2.13.4
+cocoabeans.version = 0.0.35
