+++ settings.gradle.kts
@@ -1,4 +1,3 @@
 rootProject.name = "jukebox"
-include("standalone")
-include("spigot")
-include("plugin")
+include("shared")
+include("adventure")
