+++ shared/src/main/java/com/github/idankoblik/utils/ProtocolUtils.java
