--- standalone/build.gradle.kts
@@ -1,18 +1,0 @@
-plugins {
-    id("java")
-}
-
-group = parent!!.group
-version = parent!!.version
-
-repositories {
-    mavenCentral()
-}
-
-dependencies {
-
-}
-
-tasks.test {
-    useJUnitPlatform()
-}
